Advanced Fun Fun methodes zijn ook klaar

+++ 210223_Syntra_NETCORE_opdracht/Controllers/GetallenController.cs
@@ -1,9 +1,11 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Net;
 using System.Threading.Tasks;
 
 namespace _210223_Syntra_NETCORE_opdracht.Controllers
@@ -18,32 +20,64 @@ namespace _210223_Syntra_NETCORE_opdracht.Controllers
 
         [HttpGet]
 
-        public ActionResult<int> LeesDeGetallenUitEenFile()
+        public ActionResult<int> ReadNumbersFromFile()
         {
 
 
             // If file exist, ReadAllLines and convert the data to int and add them to a int List
             if (System.IO.File.Exists(path))
             {
-                var getallenFile = System.IO.File.ReadAllLines(path);
-                getallenLijst = getallenFile.Select(int.Parse).ToList();
+                var numbersFile = System.IO.File.ReadAllLines(path);
+                getallenLijst = numbersFile.Select(int.Parse).ToList();
                 return Ok(getallenLijst);
 
             }
 
-            return NotFound();
+            return NotFound("file not found");
         }
 
 
 
-        [HttpPost("Voeg een getal toe de lijst")]
+        [HttpPost("Add a number to the list in the file")]
 
-        public ActionResult<int> VoegGetalToeDeLijst(string aText)
+        public ActionResult<int> AddNumberToListInTxt(string textNumberToInsert)
         {
-            getallenLijst.Add(Convert.ToInt16(aText));
+            if (System.IO.File.Exists(path))
+            {
 
-            
-            return Ok();
+                ArrayList numbers = new ArrayList();
+
+                StreamReader reader = new StreamReader(path);
+
+
+                string line;
+
+                while ((line = reader.ReadLine()) != null)
+
+                    numbers.Add(line);
+
+                reader.Close();
+
+                numbers.Add(textNumberToInsert);
+
+
+                StreamWriter writer = new StreamWriter(path);
+
+                foreach (string numberText in numbers)
+                {
+                    writer.WriteLine(numberText);
+
+                }
+
+
+                writer.Close();
+
+                return Ok();
+
+
+            }
+
+            return NotFound("file not found");
 
 
         }
@@ -54,18 +88,107 @@ namespace _210223_Syntra_NETCORE_opdracht.Controllers
 
         public ActionResult DeleteHetEersteGetal()
         {
-            getallenLijst.Remove(0);
+            if (System.IO.File.Exists(path))
+            {
+                getallenLijst.Remove(0);
 
-             //ik heb geen manier gevonden om de List van integeres terugzetten naar string ( To.String() wou niet werken) daarom heb ik een niuewe Lijst aangemaakt met strings
+                //ik heb geen manier gevonden om de List van integeres terugzetten naar string ( To.String() wou niet werken) daarom heb ik een niuewe Lijst aangemaakt met strings
 
-            List<string> stringGetallen = System.IO.File.ReadAllLines(path).ToList();
+                List<string> stringGetallen = System.IO.File.ReadAllLines(path).ToList();
 
-            stringGetallen.RemoveAt(0);
+                stringGetallen.RemoveAt(0);
 
-            System.IO.File.WriteAllLines(path, stringGetallen.ToArray());
+                System.IO.File.WriteAllLines(path, stringGetallen.ToArray());
 
-            return Ok();
+                return Ok();
+
+            }
+
+            return NotFound("file not found");
+
         }
+
+
+        [HttpPut]
+
+        public ActionResult InsertANewNumberAtASpecificPlace(int insertAtLineNumber, int numberToInsert)
+        {
+            if (System.IO.File.Exists(path))
+            {
+                ArrayList lines = new ArrayList();
+
+                string textToInsert = Convert.ToString(numberToInsert);
+
+                StreamReader reader = new StreamReader(path);
+
+                string line;
+
+                while ((line = reader.ReadLine()) != null)
+
+                    lines.Add(line);
+
+                reader.Close();
+
+                if (lines.Count > insertAtLineNumber - 1)
+                {
+                    lines.Insert(insertAtLineNumber - 1, textToInsert);
+                }
+
+                if (lines.Count < insertAtLineNumber)
+                {
+                    lines.Add(textToInsert);
+                }
+
+
+                StreamWriter writer = new StreamWriter(path);
+
+                foreach (string newLine in lines)
+                {
+                    writer.WriteLine(newLine);
+
+                }
+
+
+                writer.Close();
+
+                return Ok();
+            }
+
+            return NotFound("file not found");
+        }
+
+
+
+        [HttpDelete("Delete a specific number")]
+
+        public ActionResult DeleteSpecificNumber(int lineIndexNumber)
+        {
+            if (System.IO.File.Exists(path))
+            {
+                List<string> stringGetallen = System.IO.File.ReadAllLines(path).ToList();
+                int allLines = stringGetallen.Count;
+
+                if (lineIndexNumber - 1 < allLines)
+                {
+                    stringGetallen.RemoveAt(lineIndexNumber - 1);
+
+                    System.IO.File.WriteAllLines(path, stringGetallen.ToArray());
+
+                    return Ok();
+                }
+
+                return BadRequest("number out of scope");
+
+                //the other possible solution is:
+                //return StatusCode(400);
+            }
+
+
+            return NotFound("file not found");
+
+
+        }
+
